Implemented condtion ro ensure the password and confirm password supplied by the users during signup and password reset match
