+++ middleware/userAuthentication.js
@@ -47,19 +47,19 @@ const userAuthentication = async(req, res, next) => {
      }
  }
 
- const passwordRecovery = (req, res, next) => {
+//  const passwordRecovery = (req, res, next) => {
      
-        const {id} = jwt.verify(req.params.reset-passkey, PASSWORD_RECOVERY_SECRET)
+//         const {_id} = jwt.verify(req.params.token, PASSWORD_RECOVERY_SECRET)
         
-        req.user = id
+//         req.user = id
 
-        next()
+//         next()
   
- }
+//  }
 
 module.exports = {
     userAuthentication,
     emailTokenVerify,
-    passwordRecovery
+   // passwordRecovery
 
 }
