+++ routes/userRoute.js
@@ -1,6 +1,6 @@
 const express = require('express')
 const { userLogin, userSignup, verifyOTP, resendOTP, resetPasswordLink, resetPassword } = require('../controllers/userController')
-const { emailTokenVerify, passwordRecovery } = require('../middleware/userAuthentication')
+const { emailTokenVerify} = require('../middleware/userAuthentication')
 const router = express.Router()
 
 router.post('/login', userLogin)
@@ -15,9 +15,7 @@ router.get('/confirmation/:token', emailTokenVerify)
 
 router.post('/forget-password', resetPasswordLink)
 
-router.use(passwordRecovery)
-
-router.post('/reset-passkey', resetPassword)
+router.post('/reset-passkey/:token', resetPassword)
